fix: Organize tests in proper directories and fix test database setup
- Move unit tests to tests/unit/ directory
- Move integration test to tests/integration/ directory
- Fix test_context_persistence.py to create temporary database
- Add proper cleanup in finally block
- All tests now pass with correct AdCP protocol alignment
